+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/calculation/CraftingCalculatorImpl.java
@@ -8,14 +8,9 @@ import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
 
 import java.util.Collection;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import static com.refinedmods.refinedstorage.api.autocrafting.calculation.CraftingTree.root;
 
 public class CraftingCalculatorImpl implements CraftingCalculator {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CraftingCalculatorImpl.class);
-
     private final PatternRepository patternRepository;
     private final RootStorage rootStorage;
 
@@ -27,7 +22,8 @@ public class CraftingCalculatorImpl implements CraftingCalculator {
     @Override
     public <T> void calculate(final ResourceKey resource,
                               final long amount,
-                              final CraftingCalculatorListener<T> listener) {
+                              final CraftingCalculatorListener<T> listener,
+                              final CancellationToken cancellationToken) {
         CoreValidations.validateLargerThanZero(amount, "Requested amount must be greater than 0");
         final Collection<Pattern> patterns = patternRepository.getByOutput(resource);
         CraftingCalculatorListener<T> lastChildListener = null;
@@ -42,10 +38,15 @@ public class CraftingCalculatorImpl implements CraftingCalculator {
                 patternAmount
             );
             final CraftingTree<T> tree = root(pattern, rootStorage, patternAmount, patternRepository, childListener);
-            final CraftingTree.CalculationResult calculationResult = tree.calculate();
-            if (calculationResult == CraftingTree.CalculationResult.MISSING_RESOURCES) {
-                lastChildListener = childListener;
-                continue;
+            try {
+                final CraftingTree.CalculationResult calculationResult = tree.calculate(cancellationToken);
+                if (calculationResult == CraftingTree.CalculationResult.MISSING_RESOURCES) {
+                    lastChildListener = childListener;
+                    continue;
+                }
+            } catch (final CancellationException e) {
+                listener.childCalculationCancelled(childListener);
+                return;
             }
             listener.childCalculationCompleted(childListener);
             return;
@@ -54,48 +55,5 @@ public class CraftingCalculatorImpl implements CraftingCalculator {
             throw new IllegalStateException("No pattern found for " + resource);
         }
         listener.childCalculationCompleted(lastChildListener);
-    }
-
-    private boolean isCraftable(final ResourceKey resource, final long amount) {
-        final MissingResourcesCraftingCalculatorListener listener = new MissingResourcesCraftingCalculatorListener();
-        calculate(resource, amount, listener);
-        return !listener.isMissingResources();
-    }
-
-    @Override
-    public long getMaxAmount(final ResourceKey resource) {
-        try {
-            LOGGER.debug("Finding max amount for {} starting from 1", resource);
-            long low = 1;
-            long high = 1;
-            int calculationCount = 1;
-            while (isCraftable(resource, high)) {
-                low = high;
-                high = high * 2;
-                LOGGER.debug("Finding low and high for the craftable amount, currently between {} and {}", low, high);
-                calculationCount++;
-            }
-            if (low == high) {
-                return 0;
-            }
-            LOGGER.debug("Our craftable amount is between {} and {}", low, high);
-            while (low < high) {
-                final long amount = low + (high - low + 1) / 2;
-                LOGGER.debug("Trying {} (between {} and {})", amount, low, high);
-                calculationCount++;
-                if (isCraftable(resource, amount)) {
-                    LOGGER.debug("{} was craftable, increasing our low amount", amount);
-                    low = amount;
-                } else {
-                    LOGGER.debug("{} is not craftable, decreasing our high amount", amount);
-                    high = amount - 1;
-                }
-            }
-            LOGGER.debug("Found the maximum amount of {} in {} tries", low, calculationCount);
-            return low;
-        } catch (final CalculationException e) {
-            LOGGER.debug("Failed to calculate the maximum amount", e);
-            return 0;
-        }
     }
 }