+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/preview/PreviewProvider.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage.api.autocrafting.preview;
 
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CancellationToken;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
@@ -11,9 +12,11 @@ import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.4.9")
 public interface PreviewProvider {
-    CompletableFuture<Optional<Preview>> getPreview(ResourceKey resource, long amount);
+    CompletableFuture<Optional<Preview>> getPreview(ResourceKey resource, long amount,
+                                                    CancellationToken cancellationToken);
 
-    CompletableFuture<Long> getMaxAmount(ResourceKey resource);
+    CompletableFuture<Long> getMaxAmount(ResourceKey resource, CancellationToken cancellationToken);
 
-    CompletableFuture<Optional<TaskId>> startTask(ResourceKey resource, long amount, Actor actor, boolean notify);
+    CompletableFuture<Optional<TaskId>> startTask(ResourceKey resource, long amount, Actor actor, boolean notify,
+                                                  CancellationToken cancellationToken);
 }