+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/preview/PreviewType.java
@@ -7,5 +7,7 @@ public enum PreviewType {
     SUCCESS,
     MISSING_RESOURCES,
     CYCLE_DETECTED,
-    OVERFLOW
+    OVERFLOW,
+    CANCELLED,
+    NOT_AVAILABLE
 }