+++ refinedstorage-autocrafting-api/src/test/java/com/refinedmods/refinedstorage/api/autocrafting/preview/CancellationTokenImpl.java
@@ -1,0 +1,18 @@
+package com.refinedmods.refinedstorage.api.autocrafting.preview;
+
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CancellationToken;
+
+import java.util.concurrent.atomic.AtomicBoolean;
+
+public class CancellationTokenImpl implements CancellationToken {
+    private final AtomicBoolean cancelled = new AtomicBoolean(false);
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled.get();
+    }
+
+    public void cancel() {
+        cancelled.set(true);
+    }
+}