+++ refinedstorage-autocrafting-api/src/test/java/com/refinedmods/refinedstorage/api/autocrafting/task/TaskPlanTest.java
@@ -1,7 +1,9 @@
 package com.refinedmods.refinedstorage.api.autocrafting.task;
 
+import com.refinedmods.refinedstorage.api.autocrafting.CancelledCancellationToken;
 import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.autocrafting.PatternRepository;
+import com.refinedmods.refinedstorage.api.autocrafting.calculation.CancellationToken;
 import com.refinedmods.refinedstorage.api.autocrafting.calculation.CraftingCalculator;
 import com.refinedmods.refinedstorage.api.autocrafting.calculation.CraftingCalculatorImpl;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
@@ -37,7 +39,7 @@ class TaskPlanTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, OAK_PLANKS, 1);
+        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, OAK_PLANKS, 1, CancellationToken.NONE);
 
         // Assert
         assertThat(optionalPlan).isEmpty();
@@ -51,7 +53,7 @@ class TaskPlanTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, OAK_PLANKS, 1);
+        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, OAK_PLANKS, 1, CancellationToken.NONE);
 
         // Assert
         assertThat(optionalPlan).isPresent();
@@ -73,7 +75,7 @@ class TaskPlanTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, CRAFTING_TABLE, 3);
+        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, CRAFTING_TABLE, 3, CancellationToken.NONE);
 
         // Assert
         assertThat(optionalPlan).isPresent();
@@ -120,7 +122,7 @@ class TaskPlanTest {
         final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
 
         // Act
-        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, CRAFTING_TABLE, 3);
+        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, CRAFTING_TABLE, 3, CancellationToken.NONE);
 
         // Assert
         assertThat(optionalPlan).isPresent();
@@ -139,5 +141,19 @@ class TaskPlanTest {
         final Map<ResourceKey, Long> firstIngredient = craftingTableIngredients.get(0);
         assertThatThrownBy(() -> firstIngredient.put(OAK_LOG, 1L))
             .isInstanceOf(UnsupportedOperationException.class);
+    }
+
+    @Test
+    void shouldPropagateCancelCorrectly() {
+        // Arrange
+        final RootStorage storage = storage(new ResourceAmount(OAK_LOG, 1));
+        final PatternRepository patterns = patterns(OAK_PLANKS_PATTERN);
+        final CraftingCalculator sut = new CraftingCalculatorImpl(patterns, storage);
+
+        // Act
+        final Optional<TaskPlan> optionalPlan = calculatePlan(sut, OAK_PLANKS, 1, new CancelledCancellationToken());
+
+        // Assert
+        assertThat(optionalPlan).isEmpty();
     }
 }